Make the aftertouch opcode normalized

+++ src/sfizz/Defaults.h
@@ -153,8 +153,8 @@ namespace Default
     extern const OpcodeSpec<float> hiNormalized;
     extern const OpcodeSpec<float> loBipolar;
     extern const OpcodeSpec<float> hiBipolar;
-    extern const OpcodeSpec<uint8_t> loChannelAftertouch;
-    extern const OpcodeSpec<uint8_t> hiChannelAftertouch;
+    extern const OpcodeSpec<float> loChannelAftertouch;
+    extern const OpcodeSpec<float> hiChannelAftertouch;
     extern const OpcodeSpec<uint16_t> ccNumber;
     extern const OpcodeSpec<uint8_t> curveCC;
     extern const OpcodeSpec<uint8_t> smoothCC;
